refactor(currencyrate): move functionality into background-action #1021
