+++ src/common/lib/api.ts
@@ -20,8 +20,7 @@ import type {
 } from "~/types";
 
 import {
-  getAccountsCache,
-  getCurrencyRateFromCache,
+  getAccountsCache, // getCurrencyRateFromCache,
   removeAccountFromCache,
   storeAccounts,
 } from "./cache";
@@ -120,9 +119,6 @@ export const lnurlAuth = (
   options: MessageLnurlAuth["args"]
 ): Promise<LnurlAuthResponse> =>
   utils.call<LnurlAuthResponse>("lnurlAuth", options);
-
-// export const getCurrencyRate = async (currency: CURRENCIES) =>
-//   await getCurrencyRateFromCache(currency);
 
 export const getCurrencyRate = async (options: { currency: CURRENCIES }) =>
   utils.call<FixMe>("getCurrencyRate", options);
