+++ src/common/lib/cache.ts
@@ -1,8 +1,8 @@
 import dayjs from "dayjs";
 import isSameOrBefore from "dayjs/plugin/isSameOrBefore";
 import browser from "webextension-polyfill";
-import { CURRENCIES } from "~/common/constants";
-import { getFiatBtcRate } from "~/common/utils/currencyConvert";
+// import { CURRENCIES } from "~/common/constants";
+// import { getFiatBtcRate } from "~/common/utils/currencyConvert";
 import type { AccountInfo } from "~/types";
 
 dayjs.extend(isSameOrBefore);
@@ -31,41 +31,41 @@ export const removeAccountFromCache = async (id: string) => {
   return accountsCache;
 };
 
-export const getCurrencyRateFromCache = async (currency: CURRENCIES) => {
-  let currencyRateCache: { rate?: number; timestamp?: number } = {};
-  const result = await browser.storage.local.get(["currencyRate"]);
+// export const getCurrencyRateFromCache = async (currency: CURRENCIES) => {
+//   let currencyRateCache: { rate?: number; timestamp?: number } = {};
+//   const result = await browser.storage.local.get(["currencyRate"]);
 
-  if (result?.currencyRate) {
-    currencyRateCache = JSON.parse(result.currencyRate);
+//   if (result?.currencyRate) {
+//     currencyRateCache = JSON.parse(result.currencyRate);
 
-    const currentTime = dayjs();
-    const rateTimestamp = dayjs(currencyRateCache?.timestamp);
-    const rateTimestampPlusOneMinute = dayjs(rateTimestamp).add(1, "minute");
-    const isRateCurrent = currentTime.isSameOrBefore(
-      rateTimestampPlusOneMinute
-    );
+//     const currentTime = dayjs();
+//     const rateTimestamp = dayjs(currencyRateCache?.timestamp);
+//     const rateTimestampPlusOneMinute = dayjs(rateTimestamp).add(1, "minute");
+//     const isRateCurrent = currentTime.isSameOrBefore(
+//       rateTimestampPlusOneMinute
+//     );
 
-    if (isRateCurrent) {
-      return currencyRateCache.rate;
-    } else {
-      const rate = await getFiatBtcRate(currency);
-      storeCurrencyRate(rate);
-      return rate;
-    }
-  }
+//     if (isRateCurrent) {
+//       return currencyRateCache.rate;
+//     } else {
+//       const rate = await getFiatBtcRate(currency);
+//       storeCurrencyRate(rate);
+//       return rate;
+//     }
+//   }
 
-  const rate = await getFiatBtcRate(currency);
-  storeCurrencyRate(rate);
-  return rate;
-};
+//   const rate = await getFiatBtcRate(currency);
+//   storeCurrencyRate(rate);
+//   return rate;
+// };
 
-const storeCurrencyRate = (rate: number) => {
-  const currencyRate = {
-    rate,
-    timestamp: Date.now(),
-  };
+// const storeCurrencyRate = (rate: number) => {
+//   const currencyRate = {
+//     rate,
+//     timestamp: Date.now(),
+//   };
 
-  browser.storage.local.set({
-    currencyRate: JSON.stringify(currencyRate),
-  });
-};
+//   browser.storage.local.set({
+//     currencyRate: JSON.stringify(currencyRate),
+//   });
+// };
