+++ src/common/utils/currencyConvert.ts
@@ -1,7 +1,7 @@
 /**
  * Highly inspired by: https://github.com/AryanJ-NYC/bitcoin-conversion
  */
-import axios from "axios";
+// import axios from "axios";
 import { CURRENCIES } from "~/common/constants";
 import { getSettings, getCurrencyRate } from "~/common/lib/api";
 
@@ -16,33 +16,33 @@ const getCurrencySettings = async () => {
 
 const numSatsInBtc = 100_000_000;
 
-export const getFiatBtcRate = async (currency: CURRENCIES): Promise<number> => {
-  const { exchange } = await getCurrencySettings();
+// export const getFiatBtcRate = async (currency: CURRENCIES): Promise<number> => {
+//   const { exchange } = await getCurrencySettings();
 
-  let response;
+//   let response;
 
-  if (exchange === "yadio") {
-    response = await axios.get(
-      `https://api.yadio.io/exrates/${currency.toLowerCase()}`
-    );
-    const data = await response?.data;
-    return data.BTC;
-  }
+//   if (exchange === "yadio") {
+//     response = await axios.get(
+//       `https://api.yadio.io/exrates/${currency.toLowerCase()}`
+//     );
+//     const data = await response?.data;
+//     return data.BTC;
+//   }
 
-  if (exchange === "coindesk") {
-    response = await axios.get(
-      `https://api.coindesk.com/v1/bpi/currentprice/${currency.toLowerCase()}.json`
-    );
-    const data = await response?.data;
-    return data.bpi[currency].rate_float;
-  }
+//   if (exchange === "coindesk") {
+//     response = await axios.get(
+//       `https://api.coindesk.com/v1/bpi/currentprice/${currency.toLowerCase()}.json`
+//     );
+//     const data = await response?.data;
+//     return data.bpi[currency].rate_float;
+//   }
 
-  response = await axios.get(
-    `https://getalby.com/api/rates/${currency.toLowerCase()}.json`
-  );
-  const data = await response?.data;
-  return data[currency].rate_float;
-};
+//   response = await axios.get(
+//     `https://getalby.com/api/rates/${currency.toLowerCase()}.json`
+//   );
+//   const data = await response?.data;
+//   return data[currency].rate_float;
+// };
 
 const bitcoinToFiat = async (
   amountInBtc: number | string,
